refactor: 의존성 흐름을 위해  주문 테이블 단체 지정 이벤트 추상화 제거
